feat(s2n-quic-tls): client hostname verify callback
Add an option to configure a hostname verification callback on the
client side. Note that because the rust bindings don't support
configuring the callback on a per-connection basis, the verification
callback currently cannot be given the original server name that was
connected to, and must make its decision based solely on
client-creation-time context and the hostnames in the TLS certificate in
question.

In most cases where one would be implementing this callback, the
customer has custom logic for determining which hostnames to accept in
any case, so this may not be an issue, but in the future if s2n-tls rust
bindings expose this connection-specific verification call, a new trait
can be introduced to pass through the original server name, and a
blanket impl added to preserve source compatibility.

Closes: #1408