feat(s2n-quic-core): give datagram receiver access to current connection path (#1576)
This allows the receiver to differentiate the handling based on the
connection's remote address. For example, if a datagram receiver wants to
process the payload differently depending on the source address, this
allows it to do so.

The path a connection takes may change over time, so passing this state
on each on_datagram call seems like the best bet.

One alternative design (not thoroughly explored) is to expose the
datagram payloads to the Subscriber interface, and just have the
datagram provider ignore payloads.