Remove Default for datagram (Pre)ConnectionInfo structs
It doesn't make much sense for there to be default values for these, and
as more fields are added a Default may become awkward or impossible.

This removes the ability to construct these outside of s2n-quic, but
doing so is awkward if there's no reasonable default value for some of
the fields. (The next commit adds such a value for ConnectionInfo).