Switch providers to not use opaque types
The Rust compiler deals poorly with the signature of our providers,
leading to exponential compile times with additional providers. In one
real-world case this patch changes `cargo check` compile times from 6
minutes to 20 seconds, which is a significant delta. The tradeoff with
the current compiler is somewhat more verbose error messages, but given
the potentially massive wins in compile time for users of this crate
that seems worth it.