build: remove --cfg s2n_quic_unstable (#2190)
This cfg (intentionally) prevents a downstream library from wrapping
s2n-quic unstable API internally without the final build explicitly
opting into that instability. Unfortunately, with s2n-quic-dc, we are
going to be doing exactly that.

For the time being drop the cfg entirely. Since it's only set to allow
unstable features, we should technically be able to revert this change
at some point (e.g., if dc stops depending on unstable surface area): it
would be fairly similar to just breaking consumers of those unstable
features, which is already allowed.