dc: Prune peer map if we've pruned the id map (#2319)
* Prune peer list if we've pruned the path secret

This ensures that when we prune entries from the map, we bound the size
of the peer set as well.

This also updates our test coverage to include a small-size map, which
causes us to need to tweak a few assertions to account for the now
semi-random removals. I'm not very happy with the result, but I think
it's OK for now. Mid-term it probably makes sense to figure out a better
way to make sure our logic is sound (e.g., probability of removing a
recently added peer should be near zero).

* Bound requested_handshakes too