Revert "fix(s2n-quic-dc): derive crypto before opening TCP stream (#2451)"
In further testing, this introduces too much reordering under high
concurrency for our dedup tracking to mitigate. connect() can "complete"
in reverse order of starts (e.g., due to queueing of the wake events)
which leads to huge gaps (up to concurrency) in the chosen key IDs.

We can likely get the best of both worlds by deriving crypto beforehand
but only associating it with a stream just-in-time as connect()s
complete (keeping a per-peer pool just for opening connections) but
that's a more invasive change, for now just revert this.

This reverts commit 71e56fe25d7fc793ce857bdda5e16ebc6365a86b.