Replace shared map with larger bitset
This simplifies the receiver dedup implementation by replacing it with
an inline bitset, which we shift over on new arrivals. The end result is
that we have an effective tracker for ~896 packets fitting into about
the same amount of space as the previous combination of shared map and
entry data (+6 bytes if we assume 500k entries).

The new implementation is also easier to tweak, since we directly expose
the window size. It's likely that this slows down inserts a little (we
need to shift over 112 bytes) but shuffling a 112-byte array should be
sufficiently fast that I don't think it's worth worrying about it.