Rework handshakes to use a queue (#2565)
This changes our handshake strategy again. This time, I've implemented a
small simulator and run these changes in test prior to posting this...

The new approach aims to be trivially provable to handshake at the
expected rate (# of peers / handshake_period), because that's exactly
what it does. Prior attempts relied on lots of random rolls and the
law of large numbers creating that desired rate, but that doesn't always
hold (e.g., with small peer counts, we don't have large numbers), and is
harder to reason about.

This implementation, like the one in mainline, expects a p100 time 2x
the rehandshake period (due to the shuffling on each queue fill). See
comments in code for details.