Expose dc::HandshakingPath publicly
This refactors the HandshakingPath to allow consumers of the
DcPathCreated event to store a copy of the path locally, allowing
subsequent access after various events are handled (e.g., to pull out
the path secret ID for logging).

This also adjusts error propagation out of the HandshakingPath to
preserve more detail about errors, again propagating that into the
HandshakingPath for consumers to access if desired. This also avoids
panics in a few cases in favor of failing just the one handshake.