Box ControlPair to reduce copy overhead
ControlPair is 2.5kb large (2 copies of the full HMAC state). That's
maybe possible to shrink directly, but since we pass ControlPair around
by value it's good to add some indirection to avoid copying the large
value. The ControlPair is also not created at all in the TCP stream
case, so this is a pure win there.

Not measured independently (currently that's a bit slow to do) but
doesn't appear that this makes any meaningful end-to-end performance
difference. Seems like a clear win for the TCP case though.