opt(s2n-quic-dc): skip epoll registration in happy path
Currently, dcQUIC streams over TCP will be accepted, be registered with
epoll, attempt reading (usually fails), in <1ms the first data packet
arrives and we succeed reading, deregister the socket, and then hand off
the stream to the application for further reading.

We'd like to avoid the epoll registration as it uses extra CPU (even if
latency impact is minimal) so this patch uses the Linux-only
TCP_DEFER_ACCEPT to only accept sockets with data already available.
That's combined with lazy registration of sockets with Tokio's epoll by
only doing so if we get WouldBlock after attempting a read or write.

The net effect is a 8.8% (relative) drop in overall CPU usage in one of
our internal benchmarks which exercises short streams over loopback,
bringing CPU usage in the acceptor from 23% of the workload to 18%.