feat(s2n-quic-dc): Track stream write buffer allocations
This will help confirm that per-stream we only allocate a fixed number
of buffers, and provide insight into what the common buffer sizes are to
help guide tuning of allocator pool caches.

This is minimally useful for TCP streams, but is more
interesting for streams over UDP where the allocations are influenced by
retransmit returning buffers to the pool.