opt(s2n-quic-dc): Use RwLock for id/peer maps (#2799)
This uses the parking_lot RwLock which helps ensure fairness and
guarantees that writers make forward progress even with a continuous
stream of readers.

In microbenchmarks, this essentially eliminates futex syscalls on the
lock atomic, since readers are able to proceed in parallel. We likely
still have cache line contention but that's cheaper than going to sleep
on a Mutex and is a larger change to fix. The alignment to 128 bytes
should also help with the contention.