opt(s2n-quic-dc): clamp server concurrency
The lock contention referenced in the comment is a significant problem
as the number of cores increases. In the future we can consider other
mitigations for it (e.g., io_uring's fixed fd pool), but that's not
possible until we can raise our minimal kernel baseline and is a much
larger effort.