feat(s2n-quic-dc): throttle repeated successful handshakes
This mirrors prior work for *failing* handshakes by adding a throttle on
outgoing, successful handshakes to a given peer. This is particularly
relevant for workloads that happen to trigger replay detection
relatively often, which would otherwise cause us to repeatedly handshake
with the peer.

It remains generally a good idea for us to switch to new secret material
in case of e.g. spurious bitflips in memory or other issues that have
broken the old entry in some way, so retaining the
handshake-on-possible-replay makes sense, but doing so at a high rate
increases CPU utilization for little benefit.

As a drive-by improvement this also tweaks the delays to measure in
milliseconds, increasing the randomization.