opt(s2n-quic-dc): only read-lock once in happy path
This removes one of two read locks taken on the path secret map within
handshake_with_entry, which is called for each outbound stream. In
practice the double locking should be fairly cheap (just an atomic
increment most of the time), but it's still a contended cache line so
reducing operations on it should help.