Add events measuring write lock latency
This instruments acquire + total duration of write locks on the path
secret ID and address maps. These metrics should help us determine
whether these locks are impeding the *read* path (which is of higher
performance concern) without incurring as high an overhead as
instrumenting the read path directly would take.

parking_lot's RwLock will block new readers as soon as a write lock is
requested, so the total duration of the write lock is how long a reader
might wait for (plus any scheduling delay from sleeping and waking). The
acquire latency will indirectly tell us whether that latency is due to
(for some reason) slow readers or not. In general, we expect that the
durations measured here will typically be quite small.

If we find that locks are taking a long time to acquire, we can look at
replacing our global locks: sharding the table and/or replacing it
entirely with something that has per-slot locking. TBD on specifics for
that.